+++ src/main/java/net/geforcemods/securitycraft/blocks/KeyPanelBlock.java
@@ -132,7 +132,7 @@ public class KeyPanelBlock extends OwnableBlock implements IWaterLoggable {
 				activate(state, world, pos, te.getSignalLength());
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return ActionResultType.SUCCESS;