+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadChestBlock.java
@@ -104,7 +104,7 @@ public class KeypadChestBlock extends ChestBlock {
 				activate(state, world, pos, player);
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return ActionResultType.SUCCESS;