+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadDoorBlock.java
@@ -42,7 +42,7 @@ public class KeypadDoorBlock extends SpecialDoorBlock {
 				activate(state, world, pos, te.getSignalLength());
 			}
 			else if (!PlayerUtils.isHoldingItem(player, SCContent.CODEBREAKER, hand))
-				te.openPasswordGUI(player);
+				te.openPasswordGUI(world, pos, te.getOwner(), player);
 		}
 
 		return ActionResultType.SUCCESS;